feat: move Env/Hosts into sidebar, live workspace name in header (v0.4.7)
Reclaims vertical space in the API Workspace: the Env and Hosts
switchers move from a strip above the request into the left sidebar
under the Workspace selector, lifting the method+URL bar higher. The
panel header now shows the active workspace name live instead of a
static "API Workspace" label.
